+++ lib/turbo_chat.rb
@@ -2,9 +2,11 @@ require "turbo_chat/version"
 require "turbo-rails"
 require "turbo_chat/configuration"
 require "turbo_chat/model_extensions/chat_participant"
+require "turbo_chat/participant_identity"
 require "turbo_chat/permission"
 require "turbo_chat/moderation"
 require "turbo_chat/signals"
+require "turbo_chat/messages"
 require "turbo_chat/engine"
 
 module TurboChat